+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/pnl/ProfitLossPercentageCriterion.java
@@ -21,14 +21,22 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.analysis.criteria;
+package org.ta4j.core.analysis.criteria.pnl;
 
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
+import org.ta4j.core.analysis.criteria.AbstractAnalysisCriterion;
 import org.ta4j.core.num.Num;
 
-public class TotalLossCriterion extends AbstractAnalysisCriterion {
+/**
+ * Profit and loss in percentage criterion (without commissions), defined as the
+ * trade profit over the purchase price.
+ *
+ * The profit or loss in percentage over the provided {@link Trade trade(s)}.
+ * https://www.investopedia.com/ask/answers/how-do-you-calculate-percentage-gain-or-loss-investment/
+ */
+public class ProfitLossPercentageCriterion extends AbstractAnalysisCriterion {
 
     @Override
     public Num calculate(BarSeries series, TradingRecord tradingRecord) {
@@ -37,25 +45,24 @@ public class TotalLossCriterion extends AbstractAnalysisCriterion {
     }
 
     /**
-     * Calculates the gross loss of the given trade
+     * Calculates the profit or loss on a trade in percentage.
      *
      * @param series a bar series
      * @param trade  a trade
-     * @return the loss of the trade
+     * @return the profit or loss on a trade
      */
     @Override
     public Num calculate(BarSeries series, Trade trade) {
         if (trade.isClosed()) {
-            Num loss = trade.getProfit();
-            return loss.isNegative() ? loss : series.numOf(0);
-
+            Num entryPrice = trade.getEntry().getPricePerAsset();
+            return trade.getProfit().dividedBy(entryPrice).multipliedBy(series.numOf(100));
         }
         return series.numOf(0);
-
     }
 
     @Override
     public boolean betterThan(Num criterionValue1, Num criterionValue2) {
         return criterionValue1.isGreaterThan(criterionValue2);
     }
+
 }
