+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/VersusBuyAndHoldCriterionTest.java
@@ -25,6 +25,7 @@ package org.ta4j.core.analysis.criteria;
 
 import org.junit.Test;
 import org.ta4j.core.*;
+import org.ta4j.core.analysis.criteria.pnl.GrossReturnCriterion;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
@@ -46,7 +47,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(2, series),
                 Order.buyAt(3, series), Order.sellAt(5, series));
 
-        AnalysisCriterion buyAndHold = getCriterion(new TotalReturnCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new GrossReturnCriterion());
         assertNumEquals(1.10 * 1.05 / 1.05, buyAndHold.calculate(series, tradingRecord));
     }
 
@@ -56,7 +57,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(1, series),
                 Order.buyAt(2, series), Order.sellAt(5, series));
 
-        AnalysisCriterion buyAndHold = getCriterion(new TotalReturnCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new GrossReturnCriterion());
         assertNumEquals(0.95 * 0.7 / 0.7, buyAndHold.calculate(series, tradingRecord));
     }
 
@@ -65,7 +66,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest {
         MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
         Trade trade = new Trade(Order.buyAt(0, series), Order.sellAt(1, series));
 
-        AnalysisCriterion buyAndHold = getCriterion(new TotalReturnCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new GrossReturnCriterion());
         assertNumEquals((100d / 70) / (100d / 95), buyAndHold.calculate(series, trade));
     }
 
@@ -73,7 +74,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest {
     public void calculateWithNoTrades() {
         MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
 
-        AnalysisCriterion buyAndHold = getCriterion(new TotalReturnCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new GrossReturnCriterion());
         assertNumEquals(1 / 0.7, buyAndHold.calculate(series, new BaseTradingRecord()));
     }
 
@@ -102,7 +103,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest {
 
     @Test
     public void betterThan() {
-        AnalysisCriterion criterion = getCriterion(new TotalReturnCriterion());
+        AnalysisCriterion criterion = getCriterion(new GrossReturnCriterion());
         assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
         assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
     }
