+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/NetLossCriterionTest.java
@@ -21,14 +21,14 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.analysis.criteria;
+package org.ta4j.core.analysis.criteria.pnl;
 
 import org.junit.Test;
 import org.ta4j.core.AnalysisCriterion;
 import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Order;
-import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
+import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
@@ -38,68 +38,40 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
-public class TotalReturnCriterionTest extends AbstractCriterionTest {
+public class NetLossCriterionTest extends AbstractCriterionTest {
 
-    public TotalReturnCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new TotalReturnCriterion(), numFunction);
+    public NetLossCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new NetLossCriterion(), numFunction);
     }
 
     @Test
-    public void calculateWithWinningLongTrades() {
+    public void calculateOnlyWithGainTrades() {
         MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
         TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(2, series),
                 Order.buyAt(3, series), Order.sellAt(5, series));
 
-        AnalysisCriterion ret = getCriterion();
-        assertNumEquals(1.10 * 1.05, ret.calculate(series, tradingRecord));
+        AnalysisCriterion loss = getCriterion();
+        assertNumEquals(0, loss.calculate(series, tradingRecord));
     }
 
     @Test
-    public void calculateWithLosingLongTrades() {
+    public void calculateOnlyWithLossTrades() {
         MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
         TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(1, series),
                 Order.buyAt(2, series), Order.sellAt(5, series));
 
-        AnalysisCriterion ret = getCriterion();
-        assertNumEquals(0.95 * 0.7, ret.calculate(series, tradingRecord));
+        AnalysisCriterion loss = getCriterion();
+        assertNumEquals(-35, loss.calculate(series, tradingRecord));
     }
 
     @Test
-    public void calculateReturnWithWinningShortTrades() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+    public void calculateProfitWithShortTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 95, 100, 70, 80, 85, 100);
         TradingRecord tradingRecord = new BaseTradingRecord(Order.sellAt(0, series), Order.buyAt(1, series),
                 Order.sellAt(2, series), Order.buyAt(5, series));
 
-        AnalysisCriterion ret = getCriterion();
-        assertNumEquals(1.05 * 1.30, ret.calculate(series, tradingRecord));
-    }
-
-    @Test
-    public void calculateReturnWithLosingShortTrades() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 100, 80, 85, 130);
-        TradingRecord tradingRecord = new BaseTradingRecord(Order.sellAt(0, series), Order.buyAt(1, series),
-                Order.sellAt(2, series), Order.buyAt(5, series));
-
-        AnalysisCriterion ret = getCriterion();
-        assertNumEquals(0.95 * 0.70, ret.calculate(series, tradingRecord));
-    }
-
-    @Test
-    public void calculateWithNoTradesShouldReturn1() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
-
-        AnalysisCriterion ret = getCriterion();
-        assertNumEquals(1d, ret.calculate(series, new BaseTradingRecord()));
-    }
-
-    @Test
-    public void calculateWithOpenedTradeShouldReturn1() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
-        AnalysisCriterion ret = getCriterion();
-        Trade trade = new Trade();
-        assertNumEquals(1d, ret.calculate(series, trade));
-        trade.operate(0);
-        assertNumEquals(1d, ret.calculate(series, trade));
+        AnalysisCriterion loss = getCriterion();
+        assertNumEquals(-35, loss.calculate(series, tradingRecord));
     }
 
     @Test
@@ -110,7 +82,7 @@ public class TotalReturnCriterionTest extends AbstractCriterionTest {
     }
 
     @Test
-    public void testCalculateOneOpenTradeShouldReturnOne() {
-        openedTradeUtils.testCalculateOneOpenTradeShouldReturnExpectedValue(numFunction, getCriterion(), 1);
+    public void testCalculateOneOpenTradeShouldReturnZero() {
+        openedTradeUtils.testCalculateOneOpenTradeShouldReturnExpectedValue(numFunction, getCriterion(), 0);
     }
 }
